extract edittimeblock firebase functions
